% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{getL}
\alias{getL}
\title{Get the L factor of A, where A = L.D.t(L)}
\usage{
getL(p, cfm = 1:3, sparse = FALSE)
}
\description{
Get the \code{L} factor of \code{A}, with \code{sparse} option, where \code{A = L.D.t(L)}
Assumes consistent, sorted, and 1-N recoded pedigree
}

