% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{kinInv}
\alias{kinInv}
\title{Return the inverse of the Numerator Relationship Matrix}
\usage{
kinInv(p, cfm = 1:3, sparse = FALSE)
}
\description{
Given consistent pedigree, return the inverse of the Numerator Relationship Matrix either in
dense format (default) or in sparse format (\code{sparse = TRUE}).
}

