% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{kinship}
\alias{kinship}
\title{Create the Numerator Relationship Matrix with \code{sparse} option}
\usage{
kinship(p, cfm = 1:3, sparse = FALSE)
}
\description{
Create the Numerator Relationship Matrix with \code{sparse} option
}

