% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{numGen}
\alias{numGen}
\title{Get Nnmber of paternal and maternal generations in a pedigree}
\usage{
numGen(p, anc = 1, inds)
}
\description{
Get Nnmber of paternal and maternal generations in a pedigree
}

