% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{parents}
\alias{parents}
\title{Search for parents and return them for a vector of individuals}
\usage{
parents(p, i)
}
\description{
Search for parents and return them for a vector of individuals
}

