% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedBind}
\alias{pedBind}
\title{Bind two pedigree frames to form a new frame without repetition or inconsistencies}
\usage{
pedBind(x, y, xcfm = 1:3, ycfm = 1:3, Names = names(x)[xcfm],
  overlap.rm = TRUE)
}
\description{
Bind two pedigree frames to form a new frame without repetition or inconsistencies
}

