% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedCensor}
\alias{pedCensor}
\title{Check for all consistencies and fix them all}
\usage{
pedCensor(p, cfm = 1:3, only.check = FALSE, overlap.rm = TRUE)
}
\description{
Check for all consistencies and fix them all
}

