% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedEdit}
\alias{pedEdit}
\title{Edit particular inconsistencies}
\usage{
pedEdit(p, cfm = 1:3, l, duplicate.rm = TRUE, parent.record.add = TRUE,
  overlap.rm = TRUE, misid.rm = TRUE)
}
\description{
Edit particular inconsistencies
}

