% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedExtract}
\alias{pedExtract}
\title{Extract all relevant pedigree info for a subset of individuals}
\usage{
pedExtract(p.org, sample)
}
\description{
Extract all relevant pedigree info for a subset of individuals
}

