% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedMgs}
\alias{pedMgs}
\title{Create a male-only pedigree by replace all dams with their sires to get
a sire-maternal-grandsire pedigree}
\usage{
pedMgs(ped)
}
\description{
Create a male-only pedigree by replace all dams with their sires to get
a sire-maternal-grandsire pedigree
}

