% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedSort}
\alias{pedSort}
\title{Sort pedigree data in chronological order so parents appear before their progeny}
\usage{
pedSort(p.org, cfm = 1:3, recode = FALSE)
}
\arguments{
\item{p.org}{data frame or matrix with at least 3 columns for the individual and two parents.}

\item{cfm}{vector of column names or positions of individual, paternal parent and maternal parent.}

\item{recode}{logical for whether or not to output the sorted pedigree in integer-coded format.}
}
\value{
data frame for the sorted pedigree
}
\description{
\code{pedSort} receives a consistent pedigree and the column locations of the individual and
the two parents. The function then performs sorting in chronological oredr without needing
a sorting factor such as birth dates. \code{pedSort} is super fast!
}
\details{
Given an input of consistent pedigree, chronological odering is performed on the pedigree.
The input pedigree is checked first by \code{pedExamine} for all inconsistencie and an error is
generated if inconsistencies are encountered. It is recommended to always run \code{pedCensor} before
\code{pedSort}. See \sQuote{Examples} below.
}
\examples{
## Load pedigree data
data(pedigree)

## Restructure pedigree data in 'p01' and remove inconsistencies
p.clean <- pedCensor(p01)

## Sort in chronological order
p.sort <- pedSort(p.clean)

## Sort pedigree and generate integer-coded output
p.sort.recoded <- pedSort(p.org = p.clean, recode = TRUE)

}
\seealso{
\code{\link{pedCensor}}, \code{\link{pedExamine}}
}

