% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedFuns.R
\name{pedStack}
\alias{pedStack}
\title{Stack relatives in a consistent 3-column pedigree format}
\usage{
pedStack(p, v, ...)
}
\description{
Stack relatives in a consistent 3-column pedigree format
}

